[BackEnd] Email, and Order fix

+++ BackEnd/src/main/java/com/AIWebshop/AIWebshop/entity/Product.java
@@ -3,6 +3,8 @@ package com.AIWebshop.AIWebshop.entity;
 
 import jakarta.persistence.*;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "products")
 public class Product {
@@ -13,7 +15,10 @@ public class Product {
     private int id;
     private String name;
     private String description;
-    private int price;
+
+
+
+    private double price;
     private int quantity;
     @Column(name = "category_id")
     private int categoryId;
@@ -52,11 +57,11 @@ public class Product {
         this.quantity = quantity;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(double price) {
         this.price = price;
     }
